Move file watching services into OmniSharp.Services.FileWatching namespace
