+++ src/OmniSharp.Abstractions/Services/FileWatching/IFileSystemWatcher.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace OmniSharp.Services
+namespace OmniSharp.Services.FileWatching
 {
     // TODO: Flesh out this API more
     public interface IFileSystemWatcher
